test: Generalize benchmark utils
